# src/main/java/com/example/bankcards/controller/AccountController.java
package com.example.bankcards.controller;

import com.example.bankcards.dto.FIO;
import com.example.bankcards.dto.request.AccountCredentialsDTO;
import com.example.bankcards.dto.request.AccountRequestDTO;
import com.example.bankcards.dto.response.AccountResponseDTO;
import com.example.bankcards.dto.request.PhonePasswordDTO;
import com.example.bankcards.dto.response.MessageResponse;
import com.example.bankcards.mapper.AccountCredentialsMapper;
import com.example.bankcards.mapper.AccountMapper;
import com.example.bankcards.service.interfaces.AccountService;
import jakarta.validation.Valid;
import jakarta.validation.constraints.Min;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/account")
@Validated
public class AccountController {
    private final AccountService accountService;
    private final AccountCredentialsMapper accountCredentialsMapper;
    private final AccountMapper accountMapper;

    public AccountController(AccountService accountService, AccountCredentialsMapper accountCredentialsMapper, AccountMapper accountMapper) {
        this.accountService = accountService;
        this.accountCredentialsMapper = accountCredentialsMapper;
        this.accountMapper = accountMapper;
    }

    @PostMapping("/register")
    public ResponseEntity<MessageResponse> register(
            @RequestHeader(value = "X-Admin-Key", required = false) String adminKey,
            @Valid @RequestBody AccountCredentialsDTO accountCredentialsDTO
    ) throws IllegalAccessException {
        long id = accountService.registerAccount(accountCredentialsMapper.toEntity(accountCredentialsDTO), adminKey);
        return ResponseEntity.ok(new MessageResponse("Account with id " + id + " was successfully created!"));
    }

    @PostMapping("/login")
    public ResponseEntity<MessageResponse> login(
            @Valid @RequestBody PhonePasswordDTO phonePasswordDTO
    )
    {
        String token = accountService.loginAccount(phonePasswordDTO.getPhoneNumber(), phonePasswordDTO.getPassword());
        return ResponseEntity.ok(new MessageResponse(token));
    }

    @PostMapping("/create")
    @PreAuthorize("hasRole('ADMIN')")
    public ResponseEntity<MessageResponse> createUser(
            @Valid @RequestBody AccountRequestDTO accountDTO
    ) {
        accountService.createUser(accountMapper.toEntity(accountDTO), accountDTO.getCredentialsId());
        return ResponseEntity.ok(new MessageResponse("User was successfully created."));
    }

    @DeleteMapping("/{id}")
    @PreAuthorize("hasRole('ADMIN')")
    public ResponseEntity<MessageResponse> deleteUser(
            @Min(value = 0, message = "User id can't be less then 0.")
            @PathVariable("id") long id) {
        accountService.deleteUser(id);
        return ResponseEntity.ok(new MessageResponse("User and his credentials were successfully deleted."));
    }

    @GetMapping
    @PreAuthorize("hasRole('ADMIN')")
    public ResponseEntity<List<AccountResponseDTO>> getAllUsers(
            @ModelAttribute
            FIO fio,
            @Min(value = 0, message = "Page can't be less then 0.")
            @RequestParam(defaultValue = "0")
            int page,
            @Min(value = 1, message = "Size can't be less then 1.")
            @RequestParam(defaultValue = "10")
            int size
    ) {
        return ResponseEntity.status(200).body(
                accountMapper.toDTOList(accountService.getAllUsers(fio.getFirstName(), fio.getLastName(), page, size)));
    }
}